nannou::color::chromatic_adaptation
Trait AdaptFrom

pub trait AdaptFrom<S, Swp, Dwp, T>: Sized
where
    T: Component + Float,
    Swp: WhitePoint,
    Dwp: WhitePoint,
{
    // Required method
    fn adapt_from_using<M>(color: S, method: M) -> Self
       where M: TransformMatrix<Swp, Dwp, T>;

    // Provided method
    fn adapt_from(color: S) -> Self { ... }
}

Trait to convert color from one reference white point to another

Converts a color from the source white point (Swp) to the destination white point (Dwp). Uses the bradford method for conversion by default.

Required Methods

fn adapt_from_using<M>(color: S, method: M) -> Self
where M: TransformMatrix<Swp, Dwp, T>,

Convert the source color to the destination color using the specified method

Provided Methods

fn adapt_from(color: S) -> Self

Convert the source color to the destination color using the bradford method by default

Object Safety

This trait is not object safe.