nannou::color::conv
Trait IntoLinSrgba

pub trait IntoLinSrgba<S>
where
    S: Component,
{
    // Required method
    fn into_lin_srgba(self) -> Alpha<Rgb<Linear<Srgb>, S>, S>;
}

Types that may be converted directly into a linear sRGBA color representation.

This is more flexible than Into<LinSrgba<S>> as it also supports converting from different sRGBA encodings that also have different component types.

This trait is important for nannou as the Draw API works with the LinSrgba type as a target for its generic colour type API.

Required Methods

fn into_lin_srgba(self) -> Alpha<Rgb<Linear<Srgb>, S>, S>

Convert self into RGBA.