nannou::color
Trait ComponentWise

pub trait ComponentWise {
    type Scalar;

    // Required methods
    fn component_wise<F>(&self, other: &Self, f: F) -> Self
       where F: FnMut(Self::Scalar, Self::Scalar) -> Self::Scalar;
    fn component_wise_self<F>(&self, f: F) -> Self
       where F: FnMut(Self::Scalar) -> Self::Scalar;
}

Perform a unary or binary operation on each component of a color.

Required Associated Types

type Scalar

The scalar type for color components.

Required Methods

fn component_wise<F>(&self, other: &Self, f: F) -> Self
where F: FnMut(Self::Scalar, Self::Scalar) -> Self::Scalar,

Perform a binary operation on this and an other color.

fn component_wise_self<F>(&self, f: F) -> Self
where F: FnMut(Self::Scalar) -> Self::Scalar,

Perform a unary operation on this color.

Object Safety

This trait is not object safe.