nannou::draw
Module mesh

Items related to the custom mesh type used by the Draw API.

Re-exports

pub use self::builder::MeshBuilder;

pub use self::vertex::Vertex;

Modules

builder

Implementations of the lyon geometry builder traits for the Mesh.

vertex

Structs

Mesh

The custom mesh type used internally by the Draw API.

Type Aliases

Colors

Indices

MeshType

The inner mesh type used by the draw::Mesh.

Points

TexCoords