nannou::draw::mesh
Struct Mesh

pub struct Mesh { /* private fields */ }

The custom mesh type used internally by the Draw API.

Implementations

impl Mesh

pub fn raw_vertex_count(&self) -> usize

The number of raw vertices contained within the mesh.

pub fn vertex_count(&self) -> usize

The number of vertices that would be yielded by a Vertices iterator for the given mesh.

pub fn triangle_count(&self) -> usize

The number of triangles that would be yielded by a Triangles iterator for the given mesh.

pub fn points(&self) -> &[Point]

The Mesh’s vertex position channel.

pub fn indices(&self) -> &[u32]

The Mesh’s vertex indices channel.

pub fn colors(&self) -> &[Color]

The Mesh’s vertex colors channel.

pub fn tex_coords(&self) -> &[TexCoords]

The Mesh’s vertex texture coordinates channel.

pub fn push_vertex(&mut self, v: Vertex)

Push the given vertex onto the inner channels.

pub fn push_index(&mut self, i: u32)

Push the given index onto the inner Indices channel.

pub fn extend_vertices<I>(&mut self, vs: I)
where I: IntoIterator<Item = Vertex>,

Extend the mesh channels with the given vertices.

pub fn extend_indices<I>(&mut self, is: I)
where I: IntoIterator<Item = u32>,

Extend the Mesh indices channel with the given indices.

pub fn extend<V, I>(&mut self, vs: V, is: I)
where V: IntoIterator<Item = Vertex>, I: IntoIterator<Item = u32>,

Extend the Mesh with the given vertices and indices.

pub fn clear_vertices(&mut self)

Clear all vertices from the mesh.

pub fn clear_indices(&mut self)

Clear all indices from the mesh.

pub fn clear(&mut self)

Clear all vertices and indices from the mesh.

pub fn raw_vertices(&self) -> RawVertices<&Self>

Produce an iterator yielding all raw (non-index-order) vertices.

pub fn into_raw_vertices(self) -> RawVertices<Self>

Consume self and produce an iterator yielding all raw (non-index_order) vertices.

pub fn extend_from_slices( &mut self, points: &[Point], indices: &[u32], colors: &[Color], tex_coords: &[TexCoords], )

Extend the mesh from the given slices.

This is faster than extend which uses iteration internally.

**Panic!**s if the length of the given points, colors and tex_coords slices do not match.

pub fn extend_vertices_from_slices( &mut self, points: &[Point], colors: &[Color], tex_coords: &[TexCoords], )

Extend the mesh with the given slices of vertices.

pub fn extend_indices_from_slice(&mut self, indices: &[u32])

Extend the mesh with the given slices of vertices.

pub fn vertices(&self) -> Vertices<&Self>

Produce an iterator yielding all vertices in the order specified via the vertex indices.

pub fn triangles(&self) -> Triangles<&Self>

Produce an iterator yielding all triangles.

pub fn into_vertices(self) -> Vertices<Self>

Consume self and produce an iterator yielding all vertices in index-order.

pub fn into_triangles(self) -> Triangles<Self>

Consume self and produce an iterator yielding all triangles.