nannou::draw::mesh
Module vertex

Structs

IterFromColoredPoints

A type that converts an iterator yielding colored points to an iterator yielding Vertexs.

IterFromPoint2s

A type that converts an iterator yielding 2D points to an iterator yielding Vertexs.

IterFromPoints

A type that converts an iterator yielding points to an iterator yielding Vertexs.

Constants

COLOR_CHANNEL_COUNT

The number of channels in the color type.

DEFAULT_VERTEX_COLOR

Functions

default_tex_coords

Default texture coordinates, for the case where a type is not textured.

new

Simplified constructor for a draw::mesh::Vertex.

Type Aliases

Color

ColoredPoint

ColoredPoint2

Normal

Point

TexCoords

Vertex

The vertex type produced by the draw::Mesh’s inner MeshType.