nannou::draw
Module primitive

Re-exports

pub use self::arrow::Arrow;

pub use self::ellipse::Ellipse;

pub use self::line::Line;

pub use self::mesh::Mesh;

pub use self::path::Path;

pub use self::path::PathFill;

pub use self::path::PathInit;

pub use self::path::PathStroke;

pub use self::polygon::Polygon;

pub use self::polygon::PolygonInit;

pub use self::quad::Quad;

pub use self::rect::Rect;

pub use self::text::Text;

pub use self::texture::Texture;

pub use self::tri::Tri;

Modules

arrow

ellipse

line

mesh

path

polygon

quad

rect

text

texture

tri

Enums

Primitive

A wrapper around all primitive sets of properties so that they may be stored within the Draw’s drawing field while they are being drawn.