nannou::draw::primitive::ellipse
Type Alias DrawingEllipse

pub type DrawingEllipse<'a> = Drawing<'a, Ellipse>;

The drawing context for an ellipse.

Aliased Type

struct DrawingEllipse<'a> { /* private fields */ }

Implementations

impl<'a> DrawingEllipse<'a>

pub fn stroke<C>(self, color: C) -> Self
where C: IntoLinSrgba<ColorScalar>,

Stroke the outline with the given color.

pub fn radius(self, radius: f32) -> Self

Specify the width and height of the Ellipse via a given radius.

pub fn resolution(self, resolution: f32) -> Self

The number of sides used to draw the ellipse.