nannou::draw::primitive
Enum Primitive

pub enum Primitive {

    Arrow(Arrow),
    Ellipse(Ellipse),
    Line(Line),
    MeshVertexless(Vertexless),
    Mesh(Mesh),
    PathInit(PathInit),
    PathFill(PathFill),
    PathStroke(PathStroke),
    Path(Path),
    PolygonInit(PolygonInit),
    Polygon(Polygon),
    Quad(Quad),
    Rect(Rect),
    Text(Text),
    Texture(Texture),
    Tri(Tri),

}

A wrapper around all primitive sets of properties so that they may be stored within the Draw’s drawing field while they are being drawn.

This also allows us to flush all pending drawings to the mesh if Draw::to_frame is called before their respective Drawing types are dropped.

Variants

Arrow(Arrow)

Ellipse(Ellipse)

Line(Line)

MeshVertexless(Vertexless)

Mesh(Mesh)

PathInit(PathInit)

PathFill(PathFill)

PathStroke(PathStroke)

Path(Path)

PolygonInit(PolygonInit)

Polygon(Polygon)

Quad(Quad)

Rect(Rect)

Text(Text)

Texture(Texture)

Tri(Tri)