nannou::draw::primitive
Module path

Structs

Path

Properties related to drawing a Path.

PathInit

The beginning of the path building process, prior to choosing the tessellation mode (fill or stroke).

PathOptions

A path drawing context ready to specify tessellation options.

Tessellators

Mutable access to stroke and fill tessellators.

Enums

Options

Dynamically distinguish between fill and stroke tessellation options.

Traits

TessellationOptions

A set of path tessellation options (FillOptions or StrokeOptions).

Functions

points_colored_to_lyon_path

Create a lyon path for the given iterator of colored points.

points_textured_to_lyon_path

Create a lyon path for the given iterator of textured points.

Type Aliases

DrawingPath

The drawing context for a polyline whose vertices have been specified.

DrawingPathFill

The drawing context for a filled path, prior to path event submission.

DrawingPathInit

The initial drawing context for a path.

DrawingPathOptions

The drawing context for a path in the tessellation options state.

DrawingPathStroke

The drawing context for a stroked path, prior to path event submission.

PathFill

A filled path drawing context.

PathStroke

A stroked path drawing context.