nannou::draw::primitive::path
Type Alias DrawingPathInit

pub type DrawingPathInit<'a> = Drawing<'a, PathInit>;

The initial drawing context for a path.

Aliased Type

struct DrawingPathInit<'a> { /* private fields */ }

Implementations

impl<'a> DrawingPathInit<'a>

pub fn fill(self) -> DrawingPathFill<'a>

Specify that we want to use fill tessellation for the path.

The returned building context allows for specifying the fill tessellation options.

pub fn stroke(self) -> DrawingPathStroke<'a>

Specify that we want to use stroke tessellation for the path.

The returned building context allows for specifying the stroke tessellation options.