nannou::draw::primitive::path
Type Alias DrawingPathOptions

pub type DrawingPathOptions<'a, T> = Drawing<'a, PathOptions<T>>;

The drawing context for a path in the tessellation options state.

Aliased Type

struct DrawingPathOptions<'a, T> { /* private fields */ }

Implementations

impl<'a, T> DrawingPathOptions<'a, T>
where T: TessellationOptions, PathOptions<T>: Into<Primitive>, Primitive: Into<Option<PathOptions<T>>>,

pub fn events<I>(self, events: I) -> DrawingPath<'a>
where I: IntoIterator<Item = PathEvent>,

Submit the path events to be tessellated.

pub fn points<I>(self, points: I) -> DrawingPath<'a>
where I: IntoIterator, I::Item: Into<Point2>,

Submit the path events as a polyline of points.

pub fn points_closed<I>(self, points: I) -> DrawingPath<'a>
where I: IntoIterator, I::Item: Into<Point2>,

Submit the path events as a polyline of points.

An event will be generated that closes the start and end points.

pub fn points_colored<I, P, C>(self, points: I) -> DrawingPath<'a>
where I: IntoIterator<Item = (P, C)>, P: Into<Point2>, C: IntoLinSrgba<ColorScalar>,

Submit path events as a polyline of colored points.

pub fn points_colored_closed<I, P, C>(self, points: I) -> DrawingPath<'a>
where I: IntoIterator<Item = (P, C)>, P: Into<Point2>, C: IntoLinSrgba<ColorScalar>,

Submit path events as a polyline of colored points.

The path with automatically close from the end point to the start point.

pub fn points_textured<I, P, TC>( self, view: &dyn ToTextureView, points: I, ) -> DrawingPath<'a>
where I: IntoIterator<Item = (P, TC)>, P: Into<Point2>, TC: Into<TexCoords>,

Submit path events as a polyline of textured points.

pub fn points_textured_closed<I, P, TC>( self, view: &dyn ToTextureView, points: I, ) -> DrawingPath<'a>
where I: IntoIterator<Item = (P, TC)>, P: Into<Point2>, TC: Into<TexCoords>,

Submit path events as a polyline of textured points.

The path with automatically close from the end point to the start point.