nannou::draw::primitive::path
Type Alias PathFill

pub type PathFill = PathOptions<FillOptions>;

A filled path drawing context.

Aliased Type

struct PathFill { /* private fields */ }

Implementations

impl PathFill

pub fn tolerance(self, tolerance: f32) -> Self

Maximum allowed distance to the path when building an approximation.

This method is shorthand for the fill_tolerance method.

pub fn rule(self, rule: FillRule) -> Self

Specify the rule used to determine what is inside and what is outside of the shape.

Currently, only the EvenOdd rule is implemented.

This method is shorthand for the fill_rule method.