nannou::draw::primitive::polygon
Trait SetPolygon

pub trait SetPolygon: Sized {
    // Required method
    fn polygon_options_mut(&mut self) -> &mut PolygonOptions;

    // Provided methods
    fn no_fill(self) -> Self { ... }
    fn stroke_color<C>(self, color: C) -> Self
       where C: IntoLinSrgba<ColorScalar> { ... }
    fn polygon_options(self, opts: PolygonOptions) -> Self { ... }
}

A trait implemented for all polygon draw primitives.

Required Methods

fn polygon_options_mut(&mut self) -> &mut PolygonOptions

Access to the polygon builder parameters.

Provided Methods

fn no_fill(self) -> Self

Specify no fill color and in turn no fill tessellation for the polygon.

fn stroke_color<C>(self, color: C) -> Self
where C: IntoLinSrgba<ColorScalar>,

Specify a color to use for stroke tessellation.

Stroke tessellation will only be performed if this method or one of the SetStroke methods are called.

fn polygon_options(self, opts: PolygonOptions) -> Self

Specify the whole set of polygon options.

Object Safety

This trait is not object safe.