nannou::draw::primitive::polygon
Type Alias DrawingPolygonInit

pub type DrawingPolygonInit<'a> = Drawing<'a, PolygonInit>;

Initialised drawing state for a polygon.

Aliased Type

struct DrawingPolygonInit<'a> { /* private fields */ }

Implementations

impl<'a> DrawingPolygonInit<'a>

pub fn stroke<C>(self, color: C) -> Self
where C: IntoLinSrgba<ColorScalar>,

Stroke the outline with the given color.

pub fn events<I>(self, events: I) -> DrawingPolygon<'a>
where I: IntoIterator<Item = PathEvent>,

Describe the polygon with a sequence of path events.

pub fn points<I>(self, points: I) -> DrawingPolygon<'a>
where I: IntoIterator, I::Item: Into<Point2>,

Describe the polygon with a sequence of points.

pub fn points_colored<I, P, C>(self, points: I) -> DrawingPolygon<'a>
where I: IntoIterator<Item = (P, C)>, P: Into<Point2>, C: IntoLinSrgba<ColorScalar>,

Consumes an iterator of points and converts them to an iterator yielding path events.

pub fn points_textured<I, P, T>( self, view: &dyn ToTextureView, points: I, ) -> DrawingPolygon<'a>
where I: IntoIterator<Item = (P, T)>, P: Into<Point2>, T: Into<TexCoords>,

Describe the polygon with an iterator yielding textured poings.