nannou::draw::properties::spatial::dimension
Trait SetDimensions

pub trait SetDimensions: Sized {
    // Required method
    fn properties(&mut self) -> &mut Properties;

    // Provided methods
    fn width(self, w: f32) -> Self { ... }
    fn height(self, h: f32) -> Self { ... }
    fn depth(self, d: f32) -> Self { ... }
    fn w(self, w: f32) -> Self { ... }
    fn h(self, h: f32) -> Self { ... }
    fn d(self, d: f32) -> Self { ... }
    fn wh(self, v: Vec2) -> Self { ... }
    fn whd(self, v: Vec3) -> Self { ... }
    fn w_h(self, x: f32, y: f32) -> Self { ... }
    fn w_h_d(self, x: f32, y: f32, z: f32) -> Self { ... }
}

Primitives that support different dimensions.

Required Methods

fn properties(&mut self) -> &mut Properties

Provide a mutable reference to the dimension::Properties for updating.

Provided Methods

fn width(self, w: f32) -> Self

Set the absolute width for the primitive.

fn height(self, h: f32) -> Self

Set the absolute height for the primitive.

fn depth(self, d: f32) -> Self

Set the absolute depth for the primitive.

fn w(self, w: f32) -> Self

Short-hand for the width method.

fn h(self, h: f32) -> Self

Short-hand for the height method.

fn d(self, d: f32) -> Self

Short-hand for the depth method.

fn wh(self, v: Vec2) -> Self

Set the x and y dimensions for the primitive.

fn whd(self, v: Vec3) -> Self

Set the x, y and z dimensions for the primitive.

fn w_h(self, x: f32, y: f32) -> Self

Set the width and height for the primitive.

fn w_h_d(self, x: f32, y: f32, z: f32) -> Self

Set the width and height for the primitive.

Object Safety

This trait is not object safe.