nannou::draw::properties::spatial::position
Trait SetPosition

pub trait SetPosition: Sized {
    // Required method
    fn properties(&mut self) -> &mut Properties;

    // Provided methods
    fn x(self, x: f32) -> Self { ... }
    fn y(self, y: f32) -> Self { ... }
    fn z(self, z: f32) -> Self { ... }
    fn xy(self, p: Point2) -> Self { ... }
    fn xyz(self, p: Point3) -> Self { ... }
    fn x_y(self, x: f32, y: f32) -> Self { ... }
    fn x_y_z(self, x: f32, y: f32, z: f32) -> Self { ... }
}

An API for setting the position::Properties.

Required Methods

fn properties(&mut self) -> &mut Properties

Provide a mutable reference to the position::Properties for updating.

Provided Methods

fn x(self, x: f32) -> Self

Build with the given Absolute Position along the x axis.

fn y(self, y: f32) -> Self

Build with the given Absolute Position along the y axis.

fn z(self, z: f32) -> Self

Build with the given Absolute Position along the z axis.

fn xy(self, p: Point2) -> Self

Set the Position with some two-dimensional point.

fn xyz(self, p: Point3) -> Self

Set the Position with some three-dimensional point.

fn x_y(self, x: f32, y: f32) -> Self

Set the Position with x y coordinates.

fn x_y_z(self, x: f32, y: f32, z: f32) -> Self

Set the Position with x y z coordinates.

Object Safety

This trait is not object safe.