nannou::draw
Module renderer

Structs

Builder

A type aimed at simplifying construction of a draw::Renderer.

DrawError

GlyphCache

PrimitiveRender

Information about the way in which a primitive was rendered.

RenderContext

The context provided to primitives to assist with the rendering process.

Renderer

A helper type aimed at simplifying the rendering of nannou primitives via wgpu.

Scissor

The position and dimensions of the scissor.

Enums

VertexMode

A top-level indicator of whether or not

Traits

RenderPrimitive

Draw API primitives that may be rendered via the Renderer type.