nannou
Module ease

A suite of common interpolation functions often referred to as “easing” and “tweening” functions. This API is provided by the pennereq crate.

Modules

back

bounce

circ

cubic

elastic

expo

quad

quart

quint

sine

Functions

map

Maps a value from an input range to an output range

map_clamp

Maps a value from an input range to an output range clamping both the input and output to the corresponding ranges