nannou
Module event

Application, event loop and window event definitions and implementations.

Event - the default application event type.

winit::event::WindowEvent - events related to a single window.

WindowEvent - a stripped-back, simplified, newcomer-friendly version of the raw, low-level winit event.

Structs

AxisMotion

Motion along some axis of a device e.g. joystick or gamepad.

KeyboardInput

Describes a keyboard input event.

ModifiersState

Represents the current state of the keyboard modifiers

TouchEvent

The event associated with a touch at a single point.

TouchpadPressure

Pressure on a touch pad.

Update

Update event, emitted on each pass of an application loop.

Enums

ElementState

Describes the input state of a key.

Event

The default application Event type.

Key

Symbolic name for a keyboard key.

MouseButton

Describes a button of a mouse controller.

MouseScrollDelta

Describes a difference in the mouse scroll wheel state.

TouchPhase

Describes touch-screen input state.

WindowEvent

A simplified version of winit’s WindowEvent type to make it easier to get started.

Traits

LoopEvent

Event types that are compatible with the nannou app loop.