nannou::geom::path
Struct Builder

pub struct Builder { /* private fields */ }

A type used for building a 2D lyon path.

Implementations

impl Builder

pub fn new() -> Self

Begin building a new path.

pub fn with_capacity(points: usize, edges: usize) -> Self

Build a path with the given capacity for the inner path event storage.

pub fn with_svg(self) -> WithSvg<Self>

Returns a lyon builder that supports SVG commands.

pub fn flattened(self, tolerance: f32) -> Flattened<Self>

Returns a lyon builder that approximates all curves with sequences of line segments.

pub fn begin(self, to: Point2) -> Self

Sets the position in preparation for the next sub-path.

If the current sub-path contains edges, this ends the sub-path without closing it.

pub fn line_to(self, to: Point2) -> Self

Adds a line segment to the current sub-path and sets the current position.

pub fn close(self) -> Self

Closes the current sub path and sets the current position to the first position of the current sub-path.

pub fn quadratic_bezier_to(self, ctrl: Point2, to: Point2) -> Self

Add a quadratic bezier curve to the path.

pub fn cubic_bezier_to(self, ctrl1: Point2, ctrl2: Point2, to: Point2) -> Self

Add a cubic bezier curve to the path.

pub fn build(self) -> Path

Build the path and return it.

pub fn inner(&self) -> &Builder

Access to the inner lyon::path::Builder.

pub fn inner_mut(&mut self) -> &mut Builder

Mutable access to the inner lyon::path::Builder.