nannou::geom
Module quad

Structs

Quad

A quad represented by its four vertices.

Triangles

An Iterator yielding the two triangles that make up a quad.

Vertices

A simple iterator yielding each vertex in a Quad.

Constants

NUM_TRIANGLES

The number of triangles that make up a quad.

NUM_TRIANGLE_INDICES

The number of indices used to describe each triangle in the quad.

NUM_VERTICES

The number of vertices in a quad.

TRIANGLE_INDEX_TRIS

The same as triangles, but instead returns the vertex indices for each triangle.

TRIANGLE_INDICES

The same as triangles, but instead returns the vertex indices for each triangle.

Functions

centroid

Produce the centroid of the quad, aka the “mean”/“average” vertex.

triangles

Triangulates the given quad, represented by four points that describe its edges in either clockwise or anti-clockwise order.

triangles_iter

The same as triangles but provided as an Iterator.

vertices

Produce an iterator yielding each vertex in the given Quad.

Type Aliases

TrianglesIndexTris