nannou::geom::quad
Struct Quad

pub struct Quad<V = Vec3>(pub [V; 4]);

A quad represented by its four vertices.

Tuple Fields
0: [V; 4]
Implementations

impl<V> Quad<V>
where V: Vertex,

pub fn vertices(self) -> Vertices<V>

Produce an iterator yielding each vertex in the Quad.

pub fn centroid(&self) -> V
where V: Average,

Produce the centroid of the quad, aka the “mean”/“average” vertex.

pub fn triangles(&self) -> (Tri<V>, Tri<V>)

Triangulates the given quad, represented by four points that describe its edges in either clockwise or anti-clockwise order.

Example

The following rectangle

 a        b
  --------
  |      |
  |      |
  |      |
  --------
 d        c

given as

triangles([a, b, c, d])

returns

(Tri([a, b, c]), Tri([a, c, d]))

Here’s a basic code example:

use nannou::geom::{self, pt2, Quad, Tri};

fn main() {
    let a = pt2(0.0, 1.0);
    let b = pt2(1.0, 1.0);
    let c = pt2(1.0, 0.0);
    let d = pt2(0.0, 0.0);
    let quad = Quad([a, b, c, d]);
    let triangles = geom::quad::triangles(&quad);
    assert_eq!(triangles, (Tri([a, b, c]), Tri([a, c, d])));
}

pub fn triangles_iter(&self) -> Triangles<V>

The same as triangles but provided as an Iterator.

pub fn bounding_rect(self) -> Rect<<V as Vertex>::Scalar>
where V: Vertex2d,

The bounding Rect of the quad.

pub fn bounding_cuboid(self) -> Cuboid<<V as Vertex>::Scalar>
where V: Vertex3d,

The bounding Rect of the triangle.

pub fn map_vertices<F, V2>(self, map: F) -> Quad<V2>
where F: FnMut(V) -> V2,

Map the Quad’s vertices to a new type.