nannou
Module image

Items related to working with images. Currently, this module simply re-exports the renown image crate which supports reading and writing PNG, JPEG, GIF, WEBP, BMP and more.

Modules

bmpDeprecated

Decoding and Encoding of BMP Images

buffer

Iterators and other auxiliary structure for the ImageBuffer type.

codecs

Encoding and decoding for various image file formats.

ddsDeprecated

Decoding of DDS images

dxtDeprecated

Decoding of DXT (S3TC) compression

error

Contains detailed error representation.

farbfeldDeprecated

Decoding of farbfeld images

flat

Image representations for ffi.

gifDeprecated

Decoding of GIF Images

hdrDeprecated

Decoding of Radiance HDR Images

icoDeprecated

Decoding and Encoding of ICO files

imageops

Image Processing Functions

io

Input and output of images.

jpegDeprecated

Decoding and Encoding of JPEG Images

math

Mathematical helper functions and types.

pngDeprecated

Decoding and Encoding of PNG Images

pnmDeprecated

Decoding and Encoding of netpbm image formats (pbm, pgm, ppm and pam)

tgaDeprecated

Decoding and Encoding of TGA Images

tiffDeprecated

Decoding and Encoding of TIFF Images

webpDeprecated

Decoding of WebP Images

Structs

Bgr

BGR colors

Bgra

BGR colors + alpha channel

Delay

The delay of a frame relative to the previous one.

FlatSamples

A flat buffer over a (multi channel) image.

Frame

A single animation frame

Frames

An implementation dependent iterator, reading the frames as requested

ImageBuffer

Generic image buffer

Luma

Grayscale colors

LumaA

Grayscale colors + alpha channel

Pixels

Immutable pixel iterator

Progress

Represents the progress of an image operation.

Rgb

RGB colors

Rgba

RGB colors + alpha channel

SubImage

A View into another image

Enums

ColorType

An enumeration over supported color types and bit depths

DynamicImage

A Dynamic Image

ExtendedColorType

An enumeration of color types encountered in image formats.

ImageError

The generic error type for image operations.

ImageFormat

An enumeration of supported image formats. Not all formats support both encoding and decoding.

ImageOutputFormat

An enumeration of supported image formats for encoding.

Traits

AnimationDecoder

AnimationDecoder trait

EncodableLayout

Types which are safe to treat as an immutable byte slice in a pixel layout for image encoding.

GenericImage

A trait for manipulating images.

GenericImageView

Trait to inspect an image.

ImageDecoder

The trait that all decoders implement

ImageDecoderExt

Specialized image decoding not be supported by all formats

ImageEncoder

The trait all encoders implement

Pixel

A generalized pixel.

Primitive

Primitive trait from old stdlib

Functions

guess_format

Guess image format from memory block

image_dimensions

Read the dimensions of the image located at the specified path. This is faster than fully loading the image and then getting its dimensions.

load

Create a new image from a Reader

load_from_memory

Create a new image from a byte slice

load_from_memory_with_format

Create a new image from a byte slice

open

Open the image located at the path specified. The image’s format is determined from the path’s file extension.

save_buffer

Saves the supplied buffer to a file at the path specified.

save_buffer_with_format

Saves the supplied buffer to a file at the path specified in the specified format.

Type Aliases

GrayAlphaImage

Sendable grayscale + alpha channel image buffer

GrayImage

Sendable grayscale image buffer

ImageResult

Result of an image decoding/encoding process

RgbImage

Sendable Rgb image buffer

RgbaImage

Sendable Rgb + alpha channel image buffer