nannou::image::codecs::hdr
Type Alias HDRMetadata

pub type HDRMetadata = HdrMetadata;

👎Deprecated: Use HdrMetadata instead

HDR MetaData

An alias of HdrMetadata.

TODO: remove

Aliased Type

struct HDRMetadata {
    pub width: u32,
    pub height: u32,
    pub orientation: ((i8, i8), (i8, i8)),
    pub exposure: Option<f32>,
    pub color_correction: Option<(f32, f32, f32)>,
    pub pixel_aspect_ratio: Option<f32>,
    pub custom_attributes: Vec<(String, String)>,
}

Fields
width: u32

Width of decoded image. It could be either scanline length, or scanline count, depending on image orientation.

height: u32

Height of decoded image. It depends on orientation too.

orientation: ((i8, i8), (i8, i8))

Orientation matrix. For standard orientation it is ((1,0),(0,1)) - left to right, top to bottom. First pair tells how resulting pixel coordinates change along a scanline. Second pair tells how they change from one scanline to the next.

exposure: Option<f32>

Divide color values by exposure to get to get physical radiance in watts/steradian/m2

Image may not contain physical data, even if this field is set.

color_correction: Option<(f32, f32, f32)>

Divide color values by corresponding tuple member (r, g, b) to get to get physical radiance in watts/steradian/m2

Image may not contain physical data, even if this field is set.

pixel_aspect_ratio: Option<f32>

Pixel height divided by pixel width

custom_attributes: Vec<(String, String)>

All lines contained in image header are put here. Ordering of lines is preserved. Lines in the form “key=value” are represented as (“key”, “value”). All other lines are (“”, “line”)