nannou::mesh
Trait Points

pub trait Points {
    type Point;
    type Points: Channel<Element = Self::Point>;

    // Required method
    fn points(&self) -> &Self::Points;
}

All meshes must contain at least one vertex channel.

Required Associated Types

type Point

The vertex type used to represent the location of a vertex.

type Points: Channel<Element = Self::Point>

The channel type containing points.

Required Methods

fn points(&self) -> &Self::Points

Borrow the vertex channel from the mesh.