nannou::prelude::geom
Module point

Implementation of the Point types.

Currently, Point is simply a type alias for Vector. While this makes it easier for new uses to understand and switch between the two, it also conflates the the two mathematical concepts which are quite distinct. It is possible that in the future, we will switch to using distinct types. For now, we are attempting to monitor usage and feedback to determine whether or not this change is worth it.

Functions

pt2

Construct a 2-dimensional point.

pt3

Construct a 3-dimensional point.

pt4

Construct a 4-dimensional point.

Type Aliases

Point2

A 2-dimensional point type.

Point3

A 3-dimensional point type.

Point4

A 4-dimensional point type.