nannou::prelude::geom::rect
Struct Rect

pub struct Rect<S = f32> {
    pub x: Range<S>,
    pub y: Range<S>,
}

Defines a Rectangle’s bounds across the x and y axes.

Fields
x: Range<S>

The start and end positions of the Rectangle on the x axis.

y: Range<S>

The start and end positions of the Rectangle on the y axis.

Implementations

impl<S> Rect<S>
where S: Scalar + Float,

pub fn from_x_y_w_h(x: S, y: S, w: S, h: S) -> Rect<S>

Construct a Rect from the given x y coordinates and w h dimensions.

pub fn from_w_h(w: S, h: S) -> Rect<S>

Construct a Rect at origin with the given width and height.

pub fn x(&self) -> S

The position in the middle of the x bounds.

pub fn y(&self) -> S

The position in the middle of the y bounds.

pub fn x_y(&self) -> (S, S)

The centered x and y coordinates as a tuple.

pub fn x_y_w_h(&self) -> (S, S, S, S)

The Rect’s centered coordinates and dimensions in a tuple.

pub fn align_x_of(self, align: Align, other: Rect<S>) -> Rect<S>

Align self to other along the x axis in accordance with the given Align variant.

pub fn align_y_of(self, align: Align, other: Rect<S>) -> Rect<S>

Align self to other along the y axis in accordance with the given Align variant.

pub fn align_middle_x_of(self, other: Rect<S>) -> Rect<S>

Align the middle of self with the middle of the other Rect along the x axis.

pub fn align_middle_y_of(self, other: Rect<S>) -> Rect<S>

Align the middle of self with the middle of the other Rect along the y axis.

pub fn mid_top_of(self, other: Rect<S>) -> Rect<S>

Place self in the middle of the top edge of the other Rect.

pub fn mid_bottom_of(self, other: Rect<S>) -> Rect<S>

Place self in the middle of the bottom edge of the other Rect.

pub fn mid_left_of(self, other: Rect<S>) -> Rect<S>

Place self in the middle of the left edge of the other Rect.

pub fn mid_right_of(self, other: Rect<S>) -> Rect<S>

Place self in the middle of the right edge of the other Rect.

pub fn middle_of(self, other: Rect<S>) -> Rect<S>

Place self directly in the middle of the other Rect.

pub fn subdivision_ranges(&self) -> SubdivisionRanges<S>

The four ranges used for the Rect’s four subdivisions.

pub fn subdivisions(&self) -> [Rect<S>; 4]

Divide the Rect in half along the x and y axes and return the four subdivisions.

Subdivisions are yielded in the following order:

Bottom left

Bottom right

Top left

Top right

pub fn subdivisions_iter(&self) -> Subdivisions<S>

The same as subdivisions but each subdivision is yielded via the returned Iterator.

impl<S> Rect<S>
where S: Scalar,

pub fn from_corner_points(_: [S; 2], _: [S; 2]) -> Rect<S>

Construct a Rect from the coordinates of two points.

pub fn absolute(self) -> Rect<S>

Converts self to an absolute Rect so that the magnitude of each range is always positive.

pub fn overlap(self, other: Rect<S>) -> Option<Rect<S>>

The Rect representing the area in which two Rects overlap.

pub fn max(self, other: Rect<S>) -> Rect<S>
where S: Float,

The Rect that encompass the two given sets of Rect.

pub fn bottom(&self) -> S

The Rect’s lowest y value.

pub fn top(&self) -> S

The Rect’s highest y value.

pub fn left(&self) -> S

The Rect’s lowest x value.

pub fn right(&self) -> S

The Rect’s highest x value.

pub fn l_r_b_t(&self) -> (S, S, S, S)

The edges of the Rect in a tuple (left, right, bottom, top).

pub fn shift_x(self, x: S) -> Rect<S>

Shift the Rect along the x axis.

pub fn shift_y(self, y: S) -> Rect<S>

Shift the Rect along the y axis.

pub fn left_of(self, other: Rect<S>) -> Rect<S>

Align self’s right edge with the left edge of the other Rect.

pub fn right_of(self, other: Rect<S>) -> Rect<S>

Align self’s left edge with the right dge of the other Rect.

pub fn below(self, other: Rect<S>) -> Rect<S>

Align self’s top edge with the bottom edge of the other Rect.

pub fn above(self, other: Rect<S>) -> Rect<S>

Align self’s bottom edge with the top edge of the other Rect.

pub fn align_left_of(self, other: Rect<S>) -> Rect<S>

Align self’s left edge with the left edge of the other Rect.

pub fn align_right_of(self, other: Rect<S>) -> Rect<S>

Align self’s right edge with the right edge of the other Rect.

pub fn align_bottom_of(self, other: Rect<S>) -> Rect<S>

Align self’s bottom edge with the bottom edge of the other Rect.

pub fn align_top_of(self, other: Rect<S>) -> Rect<S>

Align self’s top edge with the top edge of the other Rect.

pub fn top_left_of(self, other: Rect<S>) -> Rect<S>

Place self along the top left edges of the other Rect.

pub fn top_right_of(self, other: Rect<S>) -> Rect<S>

Place self along the top right edges of the other Rect.

pub fn bottom_left_of(self, other: Rect<S>) -> Rect<S>

Place self along the bottom left edges of the other Rect.

pub fn bottom_right_of(self, other: Rect<S>) -> Rect<S>

Place self along the bottom right edges of the other Rect.

pub fn contains_point(self, _: [S; 2]) -> bool

Does the given point touch the Rectangle.

pub fn stretch_to_point(self, _: [S; 2]) -> Rect<S>

Stretches the closest edge(s) to the given point if the point lies outside of the Rect area.

pub fn closest_corner(&self, _: [S; 2]) -> Corner

Return the Corner of self that is closest to the given Point.

pub fn corners(&self) -> Quad<[S; 2]>

The four corners of the Rect.

pub fn corners_iter(&self) -> Corners<S>

An iterator yielding the four corners of the Rect.

pub fn triangles(&self) -> (Tri<[S; 2]>, Tri<[S; 2]>)

Return two Tris that represent the Rect.

pub fn triangles_iter(self) -> Triangles<[S; 2]>

An iterator yielding the Rect’s two Tri’s.

pub fn corner_at_index(&self, index: u8) -> Option<[S; 2]>

Produce the corner at the given index.

impl Rect

pub fn from_xy_wh(p: Vec2, s: Vec2) -> Rect

Construct a Rect from a given Point and Dimensions.

pub fn from_wh(s: Vec2) -> Rect

Construct a Rect at origin with the given dimensions.

pub fn from_corners(a: Vec2, b: Vec2) -> Rect

Construct a Rect from the coordinates of two points.

pub fn xy(&self) -> Vec2

The xy position in the middle of the bounds.

pub fn wh(&self) -> Vec2

The total dimensions of the Rect.

pub fn xy_wh(&self) -> (Vec2, Vec2)

Convert the Rect to a Point and Dimensions.

pub fn top_left(&self) -> Vec2

The top left corner Point.

pub fn bottom_left(&self) -> Vec2

The bottom left corner Point.

pub fn top_right(&self) -> Vec2

The top right corner Point.

pub fn bottom_right(&self) -> Vec2

The bottom right corner Point.

pub fn mid_left(&self) -> Vec2

The middle of the left edge.

pub fn mid_top(&self) -> Vec2

The middle of the top edge.

pub fn mid_right(&self) -> Vec2

The middle of the right edge.

pub fn mid_bottom(&self) -> Vec2

The middle of the bottom edge.

pub fn shift(self, v: Vec2) -> Rect

Shift the Rect by the given vector.

pub fn contains(&self, p: Vec2) -> bool

Does the given point touch the Rectangle.

pub fn stretch_to(self, p: Vec2) -> Rect

Stretches the closest edge(s) to the given point if the point lies outside of the Rect area.

impl Rect<f64>

pub fn from_xy_wh_f64(p: DVec2, s: DVec2) -> Rect<f64>

Construct a Rect from a given Point and Dimensions.

pub fn from_wh_f64(s: DVec2) -> Rect<f64>

Construct a Rect at origin with the given dimensions.

pub fn from_corners_f64(a: DVec2, b: DVec2) -> Rect<f64>

Construct a Rect from the coordinates of two points.

pub fn xy(&self) -> DVec2

The xy position in the middle of the bounds.

pub fn wh(&self) -> DVec2

The total dimensions of the Rect.

pub fn xy_wh(&self) -> (DVec2, DVec2)

Convert the Rect to a Point and Dimensions.

pub fn top_left(&self) -> DVec2

The top left corner Point.

pub fn bottom_left(&self) -> DVec2

The bottom left corner Point.

pub fn top_right(&self) -> DVec2

The top right corner Point.

pub fn bottom_right(&self) -> DVec2

The bottom right corner Point.

pub fn mid_left(&self) -> DVec2

The middle of the left edge.

pub fn mid_top(&self) -> DVec2

The middle of the top edge.

pub fn mid_right(&self) -> DVec2

The middle of the right edge.

pub fn mid_bottom(&self) -> DVec2

The middle of the bottom edge.

pub fn shift(self, v: DVec2) -> Rect<f64>

Shift the Rect by the given vector.

pub fn contains(&self, p: DVec2) -> bool

Does the given point touch the Rectangle.

pub fn stretch_to(self, p: DVec2) -> Rect<f64>

Stretches the closest edge(s) to the given point if the point lies outside of the Rect area.

impl<S> Rect<S>
where S: Scalar<Output = S> + Neg,

pub fn w(&self) -> S

The width of the Rect.

pub fn h(&self) -> S

The height of the Rect.

pub fn w_h(&self) -> (S, S)

The width and height of the Rect as a tuple.

pub fn len(&self) -> S

The length of the longest side of the rectangle.

pub fn l_t_w_h(&self) -> (S, S, S, S)

The left and top edges of the Rect along with the width and height.

pub fn l_b_w_h(&self) -> (S, S, S, S)

The left and bottom edges of the Rect along with the width and height.

pub fn pad_left(self, pad: S) -> Rect<S>

The Rect with some padding applied to the left edge.

pub fn pad_right(self, pad: S) -> Rect<S>

The Rect with some padding applied to the right edge.

pub fn pad_bottom(self, pad: S) -> Rect<S>

The rect with some padding applied to the bottom edge.

pub fn pad_top(self, pad: S) -> Rect<S>

The Rect with some padding applied to the top edge.

pub fn pad(self, pad: S) -> Rect<S>

The Rect with some padding amount applied to each edge.

pub fn padding(self, padding: Padding<S>) -> Rect<S>

The Rect with some padding applied.

pub fn relative_to_x(self, x: S) -> Rect<S>

Returns a Rect with a position relative to the given position on the x axis.

pub fn relative_to_y(self, y: S) -> Rect<S>

Returns a Rect with a position relative to the given position on the y axis.

pub fn relative_to(self, _: [S; 2]) -> Rect<S>

Returns a Rect with a position relative to the given position.

pub fn invert_x(self) -> Rect<S>

Invert the x axis (aka flip around the y axis).

pub fn invert_y(self) -> Rect<S>

Invert the y axis (aka flip around the x axis).