nannou::prelude::geom
Struct Cuboid

pub struct Cuboid<S = f32> {
    pub x: Range<S>,
    pub y: Range<S>,
    pub z: Range<S>,
}

A light-weight Cuboid type with many helper and utility methods.

The cuboid is also known as a “rectangular prism”.

Cuboid is implemented similarly to geom::Rect but with 3 axes instead of 2.

Fields
x: Range<S>

The start and end along the x axis.

y: Range<S>

The start and end along the y axis.

z: Range<S>

The start and end along the z axis.

Implementations

impl<S> Cuboid<S>
where S: Float + Scalar,

pub fn from_x_y_z_w_h_d(x: S, y: S, z: S, w: S, h: S, d: S) -> Cuboid<S>

Construct a Rect from a given centre point (x, y, z) and dimensions (width, height, depth).

pub fn x(&self) -> S

The position in the middle of the x range.

pub fn y(&self) -> S

The position in the middle of the y range.

pub fn z(&self) -> S

The position in the middle of the z range.

pub fn x_y_z(&self) -> (S, S, S)

The centered x, y and z coordinates as a tuple.

pub fn subdivision_ranges(&self) -> SubdivisionRanges<S>

The six ranges used for the Cuboid’s eight subdivisions.

pub fn x_y_z_w_h_d(&self) -> (S, S, S, S, S, S)

The position and dimensions of the cuboid.

impl<S> Cuboid<S>
where S: Scalar,

pub fn from_ranges(x: Range<S>, y: Range<S>, z: Range<S>) -> Cuboid<S>

Construct a cuboid from its x, y and z ranges.

pub fn absolute(&self) -> Cuboid<S>

Converts self to an absolute Cuboid so that the magnitude of each range is always positive.

pub fn shift_x(self, x: S) -> Cuboid<S>

Shift the cuboid along the x axis.

pub fn shift_y(self, y: S) -> Cuboid<S>

Shift the cuboid along the y axis.

pub fn shift_z(self, z: S) -> Cuboid<S>

Shift the cuboid along the z axis.

pub fn shift_by(self, _: [S; 3]) -> Cuboid<S>

Shift the cuboid by the given vector.

pub fn contains_point(&self, _: [S; 3]) -> bool

Does the given cuboid contain the given point.

pub fn stretch_to_point(self, _: [S; 3]) -> Cuboid<S>

Stretches the closest side(s) to the given point if the point lies outside of the Cuboid area.

pub fn overlap(self, other: Cuboid<S>) -> Option<Cuboid<S>>

The cuboid representing the area in which two cuboids overlap.

pub fn max(self, other: Cuboid<S>) -> Cuboid<S>
where S: Float,

The cuboid that encompass the two given cuboids.

pub fn left(&self) -> S

The start of the range along the x axis.

pub fn right(&self) -> S

The end of the range along the x axis.

pub fn bottom(&self) -> S

The start of the range along the y axis.

pub fn top(&self) -> S

The end of the range along the y axis.

pub fn front(&self) -> S

The start of the range along the z axis.

pub fn back(&self) -> S

The end of the range along the z axis.

pub fn left_quad(&self) -> Quad<[S; 3]>

The quad for the face at the start of the range along the x axis.

pub fn right_quad(&self) -> Quad<[S; 3]>

The quad for the face at the end of the range along the x axis.

pub fn bottom_quad(&self) -> Quad<[S; 3]>

The quad for the face at the start of the range along the y axis.

pub fn top_quad(&self) -> Quad<[S; 3]>

The quad for the face at the end of the range along the y axis.

pub fn front_quad(&self) -> Quad<[S; 3]>

The quad for the face at the start of the range along the z axis.

pub fn back_quad(&self) -> Quad<[S; 3]>

The quad for the face at the end of the range along the z axis.

pub fn face_quad(&self, face: Face) -> Quad<[S; 3]>

The quad for the given face.

pub fn corners(&self) -> [[S; 3]; 8]

The 8 corners of the cuboid in the following order:

y
| z
|/
0---x

  6---7
 /|  /|
2---3 |
| 4-|-5
|/  |/
0---1

pub fn corners_iter(&self) -> Corners<'_, S>

The same as corners but produces an iterator rather than a fixed-size array.

pub fn faces(&self) -> [Quad<[S; 3]>; 6]

The 6 faces of the of the cuboid in the order yielded by the Faces iterator.

pub fn faces_iter(&self) -> FaceQuads<'_, S>

An iterator yielding a quad for each face on the cuboid.

pub fn triangles_iter(&self) -> Triangles<'_, S>

Produce an iterator yielding every triangle in the cuboid (two for each face).

Uses the faces_iter method internally.

pub fn w(&self) -> S

The length of the cuboid along the x axis (aka width or w for short).

pub fn h(&self) -> S

The length of the cuboid along the y axis (aka height or h for short).

pub fn d(&self) -> S

The length of the cuboid along the z axis (aka depth or d for short).

pub fn w_h_d(&self) -> (S, S, S)

The dimensions (width, height and depth) of the cuboid as a tuple.

pub fn volume(&self) -> S

The total volume of the cuboid.

pub fn pad_left(self, pad: S) -> Cuboid<S>

The cuboid with some padding applied to the left side.

pub fn pad_right(self, pad: S) -> Cuboid<S>

The cuboid with some padding applied to the right side.

pub fn pad_bottom(self, pad: S) -> Cuboid<S>

The cuboid with some padding applied to the bottom side.

pub fn pad_top(self, pad: S) -> Cuboid<S>

The cuboid with some padding applied to the top side.

pub fn pad_front(self, pad: S) -> Cuboid<S>

The cuboid with some padding applied to the front side.

pub fn pad_back(self, pad: S) -> Cuboid<S>

The cuboid with some padding applied to the back side.

pub fn pad(self, pad: S) -> Cuboid<S>

The cuboid with some padding amount applied to each side.

impl Cuboid

pub fn from_xyz_whd(p: Vec3, s: Vec3) -> Cuboid

Construct a Rect from a given centre point (x, y, z) and dimensions (width, height, depth).

pub fn xyz(&self) -> Vec3

The xyz position in the middle of the bounds.

pub fn whd(&self) -> Vec3

The dimensions (width, height and depth) of the cuboid as a vector.

pub fn xyz_whd(&self) -> (Vec3, Vec3)

The position and dimensions of the cuboid.

pub fn shift(self, v: Vec3) -> Cuboid

Shift the cuboid by the given vector.

pub fn contains(&self, p: Vec3) -> bool

Does the given cuboid contain the given point.

pub fn stretch_to(self, p: Vec3) -> Cuboid

Stretches the closest side(s) to the given point if the point lies outside of the Cuboid area.

impl Cuboid<f64>

pub fn from_xyz_whd_f64(p: DVec3, s: DVec3) -> Cuboid<f64>

Construct a Rect from a given centre point (x, y, z) and dimensions (width, height, depth).

pub fn xyz(&self) -> DVec3

The xyz position in the middle of the bounds.

pub fn whd(&self) -> DVec3

The dimensions (width, height and depth) of the cuboid as a vector.

pub fn xyz_whd(&self) -> (DVec3, DVec3)

The position and dimensions of the cuboid.

pub fn shift(self, v: DVec3) -> Cuboid<f64>

Shift the cuboid by the given vector.

pub fn contains(&self, p: DVec3) -> bool

Does the given cuboid contain the given point.

pub fn stretch_to(self, p: DVec3) -> Cuboid<f64>

Stretches the closest side(s) to the given point if the point lies outside of the Cuboid area.