nannou::prelude::geom
Module tri

Structs

FlattenIndices

Converts an iterator yielding [usize; 3] into an iterator yielding usizes.

IterFromVertices

An iterator yielding triangles whose vertices are produced by the given iterator yielding vertices.

Tri

A triangle as three vertices.

Vertices

An iterator yielding each of the vertices of the triangle.

VerticesFromIter

An iterator that flattens an iterator yielding triangles into its vertices.

Constants

NUM_VERTICES

The number of vertices in a triangle.

Functions

flatten_index_tris

Given an iterator yielding trios of indices, produce an iterator that yields each index one at a time.

from_index_tri

Create a Tri by indexing into the given buffer.

from_vertices

Create a Tri from the next three vertices yielded by the given vertices iterator.

iter_contains

Returns the first Tri that contains the given vertex.

iter_from_vertices

Produce an iterator yielding a triangle for every three vertices yielded by the given vertices iterator.

vertices_from_iter

Produce an iterator that flattens the given iterator yielding triangles into its vertices.