nannou::text::cursor
Function xys_per_line

pub fn xys_per_line<'a, I>(
    lines_with_rects: I,
    font: &'a Font,
    text: &'a str,
    font_size: FontSize,
) -> XysPerLine<'a, I> 

Every possible cursor position within each line of text yielded by the given iterator.

Yields (xs, y_range), where y_range is the Range occupied by the line across the y axis and xs is every possible cursor position along the x axis