nannou::text
Module glyph

Logic and types specific to individual glyph layout.

Structs

Rects

An iterator yielding the Rect for each char’s Glyph in the given text.

RectsPerLine

An iterator that, for every (line, line_rect) pair yielded by the given iterator, produces an iterator that yields a Rect for every character in that line.

SelectedRects

Yields a Rect for each selected character in a single line of text.

SelectedRectsPerLine

Yields an iteraor yielding Rects for each selected character in each line of text within the given iterator yielding char Rects.

Functions

contours_to_path

Convert the given sequence of contours to a geom::Path.

index_after_cursor

Find the index of the character that directly follows the cursor at the given cursor_idx.

path_events

Produce the lyon path for the given scaled glyph.

rects_per_line

Produce an iterator that, for every (line, line_rect) pair yielded by the given iterator, produces an iterator that yields a Rect for every character in that line.

selected_rects_per_line

Produces an iterator that yields iteraors yielding Rects for each selected character in each line of text within the given iterator yielding char Rects.

Type Aliases

HalfW

The half of the width of some character (used within CharXs).

X

Some position along the X axis (used within CharXs).