nannou::text
Struct Text

pub struct Text<'a> { /* private fields */ }

An instance of some multi-line text and its layout.

Implementations

impl<'a> Text<'a>

pub fn line_infos(&self) -> &[Info]

Produce an iterator yielding information about each line.

pub fn text(&self) -> &str

The full string of text as a slice.

pub fn layout(&self) -> &Layout

The layout parameters for this text instance.

pub fn font(&self) -> &Font

The font used for this text instance.

pub fn num_lines(&self) -> usize

The number of lines in the text.

pub fn layout_rect(&self) -> Rect

The rectangle used to layout and build the text instance.

This is the same Rect that was passed to the text::Builder::build method.

pub fn bounding_rect(&self) -> Rect

The rectangle that describes the min and max bounds along each axis reached by the text.

pub fn bounding_rect_by_lines(&self) -> Rect

The rectangle that describes the min and max bounds along each axis reached by the text.

This is similar to bounding_rect but assumes that all lines have a height equal to font_size, rather than using the exact height.

pub fn width(&self) -> Scalar

The width of the widest line of text.

pub fn height(&self) -> Scalar

The exact height of the full text accounting for font size and line spacing..

pub fn height_by_lines(&self) -> Scalar

Determine the total height of a block of text with the given number of lines, font size and line_spacing (the space that separates each line of text).

The height of all lines of text are assumed to match the font_size. If looking for the exact height, see the exact_height function.

pub fn lines(&self) -> TextLines<'_>

Produce an iterator yielding each wrapped line within the Text.

pub fn line_rects(&self) -> TextLineRects<'_>

The bounding rectangle for each line.

pub fn lines_with_rects(&self) -> TextLinesWithRects<'_>

Produce an iterator yielding all lines of text alongside their bounding rects.

pub fn glyphs_per_line(&self) -> TextGlyphsPerLine<'_>

Produce an iterator yielding iterators yielding every glyph alongside its bounding rect for each line.

pub fn glyphs(&self) -> TextGlyphs<'_>

Produce an iterator yielding every glyph alongside its bounding rect.

This is the “flattened” version of the glyphs_per_line method.

pub fn path_events<'b>(&'b self) -> impl 'b + Iterator<Item = PathEvent>

Produce an iterator yielding the path events for every glyph in every line.

pub fn rt_glyphs<'b: 'a>( &'b self, window_size: Vec2, scale_factor: Scalar, ) -> impl 'a + 'b + Iterator<Item = PositionedGlyph>

Produce an iterator yielding positioned rusttype glyphs ready for caching.

The window dimensions (in logical space) and scale_factor are required to transform glyph positions into rusttype’s pixel-space, ready for caching into the rusttype glyph cache pixel buffer.

pub fn into_owned(self) -> Text<'static>

Converts this Text instance into an instance that owns the inner text string.