nannou::wgpu
Struct TextureBuilder

pub struct TextureBuilder { /* private fields */ }

A type aimed at simplifying the construction of a Texture.

The builder assumes a set of defaults describing a 128x128, non-multisampled, single-layer, non-linear sRGBA-8 texture. A suite of builder methods may be used to specify the exact properties desired.

Implementations

impl Builder

pub const REQUIRED_IMAGE_TEXTURE_USAGE: TextureUsages = wgpu::TextureUsages::COPY_DST

The minimum required texture usage when loading from an image.

pub fn from_image_view<T>(image_view: &T) -> Builder
where T: GenericImageView, <T as GenericImageView>::Pixel: Pixel,

Produce a texture descriptor from an image.

Specifically, this supports any image type implementing image::GenericImageView whose Pixel type implements Pixel.

By default, the produced builder will have the wgpu::TextureUsages returned by wgpu::TextureBuilder::default_image_texture_usage(). This is a general-purpose usage that should allow for copying to and from the texture, sampling the texture and rendering to the texture. Specifying only the texture usage required may result in better performance. It may be necessary to manually specify the the usage if STORAGE is required.

pub fn default_image_texture_usage() -> TextureUsages

The default texture usage for the case where a user has loaded a texture from an image.

impl Builder

pub const DEFAULT_SIDE: u32 = 128u32

pub const DEFAULT_DEPTH: u32 = 1u32

pub const DEFAULT_SIZE: Extent3d = _

pub const DEFAULT_ARRAY_LAYER_COUNT: u32 = 1u32

pub const DEFAULT_MIP_LEVEL_COUNT: u32 = 1u32

pub const DEFAULT_SAMPLE_COUNT: u32 = 1u32

pub const DEFAULT_DIMENSION: TextureDimension = wgpu::TextureDimension::D2

pub const DEFAULT_FORMAT: TextureFormat = wgpu::TextureFormat::Rgba8Unorm

pub const DEFAULT_USAGE: TextureUsages = _

pub const DEFAULT_DESCRIPTOR: TextureDescriptor<Option<&'static str>, &'static [TextureFormat]> = _

pub fn new() -> Builder

Creates a new Default builder

pub fn size(self, _: [u32; 2]) -> Builder

Specify the width and height of the texture.

Note: On calls to size, depth and extent the Builder will attempt to infer the wgpu::TextureDimension of its inner wgpu::TextureDescriptor by examining its size field. Use TextureBuilder::dimension() to override this behavior.

pub fn depth(self, depth: u32) -> Builder

Specify the depth of the texture.

Note: On calls to size, depth and extent the Builder will attempt to infer the wgpu::TextureDimension of its inner wgpu::TextureDescriptor by examining its size field. Use TextureBuilder::dimension() to override this behavior.

pub fn extent(self, extent: Extent3d) -> Builder

Specify the width, height and depth of the texture.

Note: On calls to size, depth and extent the Builder will attempt to infer the wgpu::TextureDimension of its inner wgpu::TextureDescriptor by examining its size field. Use TextureBuilder::dimension() to override this behavior.

pub fn dimension(self, dimension: TextureDimension) -> Builder

Specify the dimension of the texture, overriding inferred dimension.

Mainly useful for creating 2d texture arrays – override dimension with wgpu::TextureDimension::D2 on a texture with extent.depth > 1 in order to create a texture array (/ cubemap / cubemap array) instead of a 3d texture.

pub fn mip_level_count(self, count: u32) -> Builder

Specify the number of mip levels of the texture.

pub fn sample_count(self, count: u32) -> Builder

Specify the number of samples per pixel in the case that the texture is multisampled.

pub fn format(self, format: TextureFormat) -> Builder

Specify the texture format.

pub fn usage(self, usage: TextureUsages) -> Builder

Describes to the implementation how the texture is to be used.

It is important that the set of usage bits reflects the

pub fn build(self, device: &Device) -> Texture

Build the texture resulting from the specified parameters with the given device.

pub fn into_descriptor( self, ) -> TextureDescriptor<Option<&'static str>, &'static [TextureFormat]>

Consumes the builder and returns the resulting wgpu::TextureDescriptor.