nannou::wgpu
Type Alias DeviceDescriptor

pub type DeviceDescriptor<'a> = DeviceDescriptor<Option<&'a str>>;

Describes a Device.

For use with Adapter::request_device.

Corresponds to WebGPU GPUDeviceDescriptor.

Aliased Type

struct DeviceDescriptor<'a> {
    pub label: Option<&'a str>,
    pub features: Features,
    pub limits: Limits,
}

Fields
label: Option<&'a str>

Debug label for the device.

features: Features

Features that the device should support. If any feature is not supported by the adapter, creating a device will panic.

limits: Limits

Limits that the device should support. If any limit is “better” than the limit exposed by the adapter, creating a device will panic.