nannou
Module window

The nannou Window API.

Create a new window via app.new_window(). This produces a Builder which can be used to build a Window.

Structs

Builder

A context for building a window.

Id

Identifier of a window. Unique for each window.

SurfaceConfigurationBuilder

Surface configuration for which nannou will provide a default if unspecified.

Window

A nannou window.

Enums

BuildError

Errors that might occur while building the window.

Fullscreen

Fullscreen modes.

Constants

DEFAULT_DIMENSIONS

The default dimensions used for a window in the case that none are specified.

MIN_SC_PIXELS

The default minimum dimensions used for the surface

Type Aliases

ClosedFn

A function for processing window closed events.

DroppedFileFn

A function for processing dropped file events.

EventFn

A function for processing window events.

FocusedFn

A function for processing window focused events.

HoveredFileCancelledFn

A function for processing hovered file cancelled events.

HoveredFileFn

A function for processing hovered file events.

KeyPressedFn

A function for processing key press events.

KeyReleasedFn

A function for processing key release events.

MouseEnteredFn

A function for processing mouse entered events.

MouseExitedFn

A function for processing mouse exited events.

MouseMovedFn

A function for processing mouse moved events.

MousePressedFn

A function for processing mouse pressed events.

MouseReleasedFn

A function for processing mouse released events.

MouseWheelFn

A function for processing mouse wheel events.

MovedFn

A function for processing window moved events.

RawEventFn

A function for processing raw winit window events.

RawViewFn

The user function type for drawing their model to the surface of a single window.

ReceivedCharacterFn

A function for processing received characters.

ResizedFn

A function for processing window resized events.

SketchFn

The same as ViewFn, but provides no user model to draw from.

TouchFn

A function for processing touch events.

TouchpadPressureFn

A function for processing touchpad pressure events.

UnfocusedFn

A function for processing window unfocused events.

ViewFn

The user function type for drawing their model to the surface of a single window.